@@ -3,11 +3,9 @@
 namespace App\Controller;
 
 use App\Action\Copy\InsertAction;
-use App\Action\Copy\InsertUserroomAction;
 use App\Action\Copy\RemoveAction;
 use App\Services\CopyService;
 use App\Services\LegacyEnvironment;
-use App\Utils\RoomService;
 use cs_item;
 use cs_room_item;
 use Exception;
@@ -27,15 +25,24 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
  */
 class CopyController extends BaseController
 {
-    protected $roomService;
+    /**
+     * @var CopyService
+     */
+    private CopyService $copyService;
 
+    /**
+     * @required
+     * @param CopyService $copyService
+     */
+    public function setCopyService(CopyService $copyService): void
+    {
+        $this->copyService = $copyService;
+    }
 
     /**
      * @Route("/room/{roomId}/copy/feed/{start}/{sort}")
      * @Template()
      * @param Request $request
-     * @param CopyService $copyService
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @param int $max
      * @param int $start
@@ -44,8 +51,6 @@ class CopyController extends BaseController
      */
     public function feedAction(
         Request $request,
-        CopyService $copyService,
-        LegacyEnvironment $environment,
         int $roomId,
         int $max = 10,
         int $start = 0,
@@ -57,8 +62,8 @@ class CopyController extends BaseController
         if (!$copyFilter) {
             $copyFilter = $request->query->get('copy_filter');
         }
-        
-        $roomItem = $this->loadRoom($environment, $roomId);
+
+        $roomItem = $this->loadRoom($roomId);
 
         if ($roomItem->isPrivateRoom()) {
             $rubrics = [
@@ -76,16 +81,16 @@ class CopyController extends BaseController
         if ($copyFilter) {
             // setup filter form
             $filterForm = $this->createFilterForm($roomItem);
-    
+
             // manually bind values from the request
             $filterForm->submit($copyFilter);
-    
+
             // apply filter
-            $copyService->setFilterConditions($filterForm);
+            $this->copyService->setFilterConditions($filterForm);
         }
 
         // get announcement list from manager service 
-        $entries = $copyService->getListEntries($roomId, $max, $start, $sort);
+        $entries = $this->copyService->getListEntries($roomId, $max, $start, $sort);
 
         $stackRubrics = ['date', 'material', 'discussion', 'todo'];
 
@@ -111,30 +116,26 @@ class CopyController extends BaseController
      * @Route("/room/{roomId}/copy")
      * @Template()
      * @param Request $request
-     * @param CopyService $copyService
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @return array
      */
     public function listAction(
         Request $request,
-        CopyService $copyService,
-        LegacyEnvironment $environment,
         int $roomId
     ) {
-        $roomItem = $this->loadRoom($environment, $roomId);
+        $roomItem = $this->loadRoom($roomId);
         $filterForm = $this->createFilterForm($roomItem);
 
         // apply filter
         $filterForm->handleRequest($request);
         if ($filterForm->isSubmitted() && $filterForm->isValid()) {
             // set filter conditions
-            $copyService->setFilterConditions($filterForm);
+            $this->copyService->setFilterConditions($filterForm);
         }
 
         // get number of items
-        $itemsCountArray = $copyService->getCountArray($roomId);
-        
+        $itemsCountArray = $this->copyService->getCountArray($roomId);
+
         return [
             'roomId' => $roomId,
             'form' => $filterForm->createView(),
@@ -167,59 +168,11 @@ class CopyController extends BaseController
         return $action->execute($room, $items);
     }
 
-//    public function xhrInsertStackAction($roomId, Request $request)
-//    {
-//        $privateRoomItem = $legacyEnvironment->getCurrentUser()->getOwnRoom();
-//        $legacyEnvironment->changeContextToPrivateRoom($privateRoomItem->getItemID());
-//
-//        $errorArray = [];
-//        if (!empty($selectedIds)) {
-//            foreach ($selectedIds as $id) {
-//
-//                // get item to copy
-//                $item = $itemService->getItem($id);
-//
-//                // for now, we only copy materials, dates, discussions and todos
-//                if (in_array($item->getItemType(), array(CS_MATERIAL_TYPE, CS_DATE_TYPE, CS_DISCUSSION_TYPE, CS_TODO_TYPE))) {
-//
-//                    // archive
-//                    $toggleArchive = false;
-//                    if ($item->isArchived() and !$legacyEnvironment->isArchiveMode()) {
-//                        $toggleArchive = true;
-//                        $legacyEnvironment->toggleArchiveMode();
-//                    }
-//
-//                    // archive
-//                    $importItem = $itemService->getTypedItem($id);
-//
-//                    // archive
-//                    if ($toggleArchive) {
-//                        $legacyEnvironment->toggleArchiveMode();
-//                    }
-//
-//                    // archive
-//                    $copy = $importItem->copy();
-//
-//                    $err = $copy->getErrorArray();
-//                    if (!empty($err)) {
-//                        $errorArray[$copy->getItemID() ] = $err;
-//                    }
-//                }
-//            }
-//        }
-//
-//        if (!empty($errorArray)) {
-//            $message = '<i class=\'uk-icon-justify uk-icon-medium uk-icon-check-bolt\'></i> '.implode(', ', $errorArray);
-//        } else {
-//            $message = '<i class=\'uk-icon-justify uk-icon-medium uk-icon-copy\'></i> '.$translator->transChoice('inserted %count% entries in my stack',count($selectedIds), array('%count%' => count($selectedIds)));
-//        }
-//    }
-
     /**
      * @Route("/room/{roomId}/copy/xhr/remove", condition="request.isXmlHttpRequest()")
      * @param Request $request
      * @param int $roomId
-     * @return
+     * @return mixed
      * @throws Exception
      */
     public function xhrRemoveAction(
@@ -246,7 +199,6 @@ class CopyController extends BaseController
         $selectAll,
         $itemIds = []
     ) {
-        $copyService = $this->get('commsy.copy_service');
 
         if ($selectAll) {
             if ($request->query->has('copy_filter')) {
@@ -257,12 +209,12 @@ class CopyController extends BaseController
                 $filterForm->submit($currentFilter);
 
                 // apply filter
-                $copyService->setFilterConditions($filterForm);
+                $this->copyService->setFilterConditions($filterForm);
             }
 
-            return $copyService->getListEntries($roomItem->getItemID());
+            return $this->copyService->getListEntries($roomItem->getItemID());
         } else {
-            return $copyService->getCopiesById($roomItem->getItemID(), $itemIds);
+            return $this->copyService->getCopiesById($roomItem->getItemID(), $itemIds);
         }
     }
 
@@ -295,16 +247,14 @@ class CopyController extends BaseController
     }
 
     private function loadRoom(
-        LegacyEnvironment $environment,
         int $roomId
     ) {
-        $legacyEnvironment = $environment->getEnvironment();
 
-        $roomManager = $legacyEnvironment->getRoomManager();
+        $roomManager = $this->legacyEnvironment->getRoomManager();
         $roomItem = $roomManager->getItem($roomId);
 
         if (!$roomItem) {
-            $privateRoomManager = $legacyEnvironment->getPrivateRoomManager();
+            $privateRoomManager = $this->legacyEnvironment->getPrivateRoomManager();
             $roomItem = $privateRoomManager->getItem($roomId);
 
             if (!$roomItem) {
