@@ -554,7 +554,8 @@ class cs_announcement_manager extends cs_manager implements cs_export_import_int
 
   function delete ($item_id) {
      $current_datetime = getCurrentDateTimeInMySQL();
-     $user_id = $this->_current_user->getItemID();
+     $user = $this->_environment->getCurrentUser();
+     $user_id = $user->getItemID();
      $query = 'UPDATE '.$this->addDatabasePrefix('announcement').' SET '.
               'deletion_date="'.$current_datetime.'",'.
               'deleter_id="'.encode(AS_DB,$user_id).'"'.
