@@ -416,7 +416,7 @@ class cs_step_manager extends cs_manager implements cs_export_import_interface {
               'modification_date="'.$current_datetime.'",'.
               'title="'.encode(AS_DB,$item->getTitle()).'",'.
               'description="'.encode(AS_DB,$item->getDescription()).'",'.
-              'minutes="'.encode(AS_DB,$item->getMinutes()).'",'.
+               $this->returnQuerySentenceIfFieldIsValid($item->getMinutes(), "minutes").
               'time_type="'.encode(AS_DB,$item->getTimeType()).'",'.
               'todo_item_id="'.encode(AS_DB,$item->getTodoID()).'"';
      $result = $this->_db_connector->performQuery($query);
