@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Action\Copy\CopyAction;
+use App\Action\Delete\DeleteAction;
 use App\Action\Delete\DeleteDate;
 use App\Action\Download\DownloadAction;
 use App\Entity\Calendars;
@@ -13,7 +14,6 @@ use App\Form\Type\AnnotationType;
 use App\Form\Type\DateImportType;
 use App\Form\Type\DateType;
 use App\Services\CalendarsService;
-use App\Services\LegacyEnvironment;
 use App\Services\LegacyMarkup;
 use App\Services\PrintService;
 use App\Utils\AnnotationService;
@@ -46,12 +46,26 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class DateController extends BaseController
 {
+
+    /**
+     * @var DateService
+     */
+    private DateService $dateService;
+
+    /**
+     * @required
+     * @param DateService $dateService
+     */
+    public function setDateService(DateService $dateService): void
+    {
+        $this->dateService = $dateService;
+    }
+
+
     /**
      * @Route("/room/{roomId}/date/feed/{start}/{sort}")
      * @Template()
      * @param Request $request
-     * @param DateService $dateService
-     * @param ReaderService $readerService
      * @param int $roomId
      * @param int $max
      * @param int $start
@@ -60,8 +74,6 @@ class DateController extends BaseController
      */
     public function feedAction(
         Request $request,
-        DateService $dateService,
-        ReaderService $readerService,
         int $roomId,
         int $max = 10,
         int $start = 0,
@@ -82,10 +94,10 @@ class DateController extends BaseController
             // manually bind values from the request
             $filterForm->submit($dateFilter);
             // set filter conditions on the date manager
-            $dateService->setFilterConditions($filterForm);
+            $this->dateService->setFilterConditions($filterForm);
         } else {
-            $dateService->setPastFilter(false);
-            $dateService->hideDeactivatedEntries();
+            $this->dateService->setPastFilter(false);
+            $this->dateService->hideDeactivatedEntries();
         }
 
         // Correct sort from "date" to "time". Applies only in date rubric.
@@ -96,14 +108,14 @@ class DateController extends BaseController
         }
 
         // get material list from manager service
-        $dates = $dateService->getListDates($roomId, $max, $start, $sort);
+        $dates = $this->dateService->getListDates($roomId, $max, $start, $sort);
 
         $this->get('session')->set('sortDates', $sort);
 
         $readerList = array();
         $allowedActions = array();
         foreach ($dates as $item) {
-            $readerList[$item->getItemId()] = $readerService->getChangeStatus($item->getItemId());
+            $readerList[$item->getItemId()] = $this->readerService->getChangeStatus($item->getItemId());
             if ($this->isGranted('ITEM_EDIT', $item->getItemID())) {
                 $allowedActions[$item->getItemID()] = array('markread', 'copy', 'save', 'delete');
             } else {
@@ -123,18 +135,13 @@ class DateController extends BaseController
      * @Route("/room/{roomId}/date")
      * @Template()
      * @param Request $request
-     * @param DateService $dateService
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @return array
      */
     public function listAction(
         Request $request,
-        DateService $dateService,
-        LegacyEnvironment $environment,
         int $roomId
     ) {
-        $legacyEnvironment = $environment->getEnvironment();
         $roomItem = $this->getRoom($roomId);
 
         $filterForm = $this->createFilterForm($roomItem);
@@ -143,13 +150,13 @@ class DateController extends BaseController
         $filterForm->handleRequest($request);
         if ($filterForm->isSubmitted() && $filterForm->isValid()) {
             // set filter conditions in date manager
-            $dateService->setFilterConditions($filterForm);
+            $this->dateService->setFilterConditions($filterForm);
         } else {
-            $dateService->setPastFilter(false);
-            $dateService->hideDeactivatedEntries();
+            $this->dateService->setPastFilter(false);
+            $this->dateService->hideDeactivatedEntries();
         }
 
-        $itemsCountArray = $dateService->getCountArray($roomId);
+        $itemsCountArray = $this->dateService->getCountArray($roomId);
 
         $usageInfo = false;
         /** @noinspection PhpUndefinedMethodInspection */
@@ -174,12 +181,12 @@ class DateController extends BaseController
         if ($roomItem->isOpenForGuests()) {
             $iCal['show'] = true;
         } else {
-            $currentUserItem = $legacyEnvironment->getCurrentUserItem();
+            $currentUserItem = $this->legacyEnvironment->getCurrentUserItem();
 
             if ($currentUserItem->isUser()) {
                 $iCal['show'] = true;
 
-                $hashManager = $legacyEnvironment->getHashManager();
+                $hashManager = $this->legacyEnvironment->getHashManager();
                 $iCalHash = $hashManager->getICalHashForUser($currentUserItem->getItemID());
 
                 $iCal['aboUrl'] = $this->generateUrl('app_ical_getcontent', [
@@ -208,55 +215,51 @@ class DateController extends BaseController
             'iCal' => $iCal,
             'calendars' => $calendars,
             'isArchived' => $roomItem->isArchived(),
-            'user' => $legacyEnvironment->getCurrentUserItem(),
+            'user' => $this->legacyEnvironment->getCurrentUserItem(),
         ];
     }
 
     /**
      * @Route("/room/{roomId}/date/print/{sort}", defaults={"sort" = "none"})
      * @param Request $request
-     * @param DateService $dateService
      * @param PrintService $printService
-     * @param ReaderService $readerService
      * @param int $roomId
      * @param string $sort
      * @return Response
      */
     public function printlistAction(
         Request $request,
-        DateService $dateService,
         PrintService $printService,
-        ReaderService $readerService,
         int $roomId,
         string $sort
     ) {
         $roomItem = $this->getRoom($roomId);
         $filterForm = $this->createFilterForm($roomItem);
-        $numAllDates = $dateService->getCountArray($roomId)['countAll'];
+        $numAllDates = $this->dateService->getCountArray($roomId)['countAll'];
 
         // apply filter
         $filterForm->handleRequest($request);
         if ($filterForm->isSubmitted() && $filterForm->isValid()) {
             // set filter conditions in date manager
-            $dateService->setFilterConditions($filterForm);
+            $this->dateService->setFilterConditions($filterForm);
         } else {
-            $dateService->setPastFilter(false);
+            $this->dateService->setPastFilter(false);
         }
 
         // get date list from manager service
         if ($sort != "none") {
-            $dates = $dateService->getListDates($roomId, $numAllDates, 0, $sort);
+            $dates = $this->dateService->getListDates($roomId, $numAllDates, 0, $sort);
         }
         elseif ($this->get('session')->get('sortDates')) {
-            $dates = $dateService->getListDates($roomId, $numAllDates, 0, $this->get('session')->get('sortDates'));
+            $dates = $this->dateService->getListDates($roomId, $numAllDates, 0, $this->get('session')->get('sortDates'));
         }
         else {
-            $dates = $dateService->getListDates($roomId, $numAllDates, 0, 'date');
+            $dates = $this->dateService->getListDates($roomId, $numAllDates, 0, 'date');
         }
 
         $readerList = array();
         foreach ($dates as $item) {
-            $reader = $readerService->getLatestReader($item->getItemId());
+            $reader = $this->readerService->getLatestReader($item->getItemId());
             if ( empty($reader) ) {
                $readerList[$item->getItemId()] = 'new';
             } elseif ( $reader['read_date'] < $item->getModificationDate() ) {
@@ -264,7 +267,7 @@ class DateController extends BaseController
             }
         }
 
-        $itemsCountArray = $dateService->getCountArray($roomId);
+        $itemsCountArray = $this->dateService->getCountArray($roomId);
 
         $html = $this->renderView('date/list_print.html.twig', [
             'roomId' => $roomId,
@@ -282,26 +285,22 @@ class DateController extends BaseController
      * @Template()
      * @param Request $request
      * @param DateService $dateService
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @return array
      */
     public function calendarAction(
         Request $request,
-        DateService $dateService,
-        LegacyEnvironment $environment,
         int $roomId
     ) {
-        $legacyEnvironment = $environment->getEnvironment();
         $roomItem = $this->getRoom($roomId);
         $filterForm = $this->createFilterForm($roomItem, false);
 
         // apply filter
         $filterForm->handleRequest($request);
         if ($filterForm->isSubmitted() && $filterForm->isValid()) {
-            $dateService->setFilterConditions($filterForm);
+            $this->dateService->setFilterConditions($filterForm);
         } else {
-            $dateService->setPastFilter(false);
+            $this->dateService->setPastFilter(false);
         }
 
         $usageInfo = false;
@@ -327,12 +326,12 @@ class DateController extends BaseController
         if ($roomItem->isOpenForGuests()) {
             $iCal['show'] = true;
         } else {
-            $currentUserItem = $legacyEnvironment->getCurrentUserItem();
+            $currentUserItem = $this->legacyEnvironment->getCurrentUserItem();
 
             if ($currentUserItem->isUser()) {
                 $iCal['show'] = true;
 
-                $hashManager = $legacyEnvironment->getHashManager();
+                $hashManager = $this->legacyEnvironment->getHashManager();
                 $iCalHash = $hashManager->getICalHashForUser($currentUserItem->getItemID());
 
                 $iCal['aboUrl'] = $this->generateUrl('app_ical_getcontent', [
@@ -387,13 +386,8 @@ class DateController extends BaseController
      * @param Request $request
      * @param AnnotationService $annotationService
      * @param CategoryService $categoryService
-     * @param DateService $dateService
-     * @param ItemService $itemService
-     * @param ReaderService $readerService
      * @param TopicService $topicService
-     * @param TranslatorInterface $translator
      * @param LegacyMarkup $legacyMarkup
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @param int $itemId
      * @return array
@@ -402,27 +396,21 @@ class DateController extends BaseController
         Request $request,
         AnnotationService $annotationService,
         CategoryService $categoryService,
-        DateService $dateService,
-        ItemService $itemService,
-        ReaderService $readerService,
         TopicService $topicService,
-        TranslatorInterface $translator,
         LegacyMarkup $legacyMarkup,
-        LegacyEnvironment $environment,
         int $roomId,
         int $itemId
     ) {
-        $date = $dateService->getDate($itemId);
+        $date = $this->dateService->getDate($itemId);
 
-        $legacyEnvironment = $environment->getEnvironment();
         $item = $date;
-        $reader_manager = $legacyEnvironment->getReaderManager();
+        $reader_manager = $this->legacyEnvironment->getReaderManager();
         $reader = $reader_manager->getLatestReader($item->getItemID());
         if(empty($reader) || $reader['read_date'] < $item->getModificationDate()) {
             $reader_manager->markRead($item->getItemID(), $item->getVersionID());
         }
 
-        $noticed_manager = $legacyEnvironment->getNoticedManager();
+        $noticed_manager = $this->legacyEnvironment->getNoticedManager();
         $noticed = $noticed_manager->getLatestNoticed($item->getItemID());
         if(empty($noticed) || $noticed['read_date'] < $item->getModificationDate()) {
             $noticed_manager->markNoticed($item->getItemID(), $item->getVersionID());
@@ -434,12 +422,12 @@ class DateController extends BaseController
 
         $itemArray = array($date);
 
-        $current_context = $legacyEnvironment->getCurrentContextItem();
+        $current_context = $this->legacyEnvironment->getCurrentContextItem();
 
-        $readerManager = $legacyEnvironment->getReaderManager();
+        $readerManager = $this->legacyEnvironment->getReaderManager();
 
-        $userManager = $legacyEnvironment->getUserManager();
-        $userManager->setContextLimit($legacyEnvironment->getCurrentContextID());
+        $userManager = $this->legacyEnvironment->getUserManager();
+        $userManager->setContextLimit($this->legacyEnvironment->getCurrentContextID());
         $userManager->setUserLimit();
         $userManager->select();
         $user_list = $userManager->get();
@@ -472,14 +460,14 @@ class DateController extends BaseController
         $readerList = array();
         $modifierList = array();
         foreach ($itemArray as $item) {
-            $reader = $readerService->getLatestReader($item->getItemId());
+            $reader = $this->readerService->getLatestReader($item->getItemId());
             if ( empty($reader) ) {
                $readerList[$item->getItemId()] = 'new';
             } elseif ( $reader['read_date'] < $item->getModificationDate() ) {
                $readerList[$item->getItemId()] = 'changed';
             }
             
-            $modifierList[$item->getItemId()] = $itemService->getAdditionalEditorsForItem($item);
+            $modifierList[$item->getItemId()] = $this->itemService->getAdditionalEditorsForItem($item);
         }
 
         // annotation form
@@ -493,12 +481,12 @@ class DateController extends BaseController
         }
 
         $alert = null;
-        if ($dateService->getDate($itemId)->isLocked()) {
+        if ($this->dateService->getDate($itemId)->isLocked()) {
             $alert['type'] = 'warning';
-            $alert['content'] = $translator->trans('item is locked', array(), 'item');
+            $alert['content'] = $this->translator->trans('item is locked', array(), 'item');
         } else if ($date->isExternal()) {
             $alert['type'] = 'warning';
-            $alert['content'] = $translator->trans('date is external', array(), 'date');
+            $alert['content'] = $this->translator->trans('date is external', array(), 'date');
         }
 
         $pathTopicItem = null;
@@ -506,29 +494,29 @@ class DateController extends BaseController
             $pathTopicItem = $topicService->getTopic($request->query->get('path'));
         }
 
-        $legacyMarkup->addFiles($itemService->getItemFileList($itemId));
-        $amountAnnotations = $annotationService->getListAnnotations($roomId, $dateService->getDate($itemId)->getItemId(), null, null);
+        $legacyMarkup->addFiles($this->itemService->getItemFileList($itemId));
+        $amountAnnotations = $annotationService->getListAnnotations($roomId, $this->dateService->getDate($itemId)->getItemId(), null, null);
 
         return array(
             'roomId' => $roomId,
-            'date' => $dateService->getDate($itemId),
+            'date' => $this->dateService->getDate($itemId),
             'amountAnnotations' => sizeof($amountAnnotations),
             'readerList' => $readerList,
             'modifierList' => $modifierList,
-            'user' => $legacyEnvironment->getCurrentUserItem(),
+            'user' => $this->legacyEnvironment->getCurrentUserItem(),
             'annotationForm' => $form->createView(),
             'userCount' => $all_user_count,
             'readCount' => $read_count,
             'readSinceModificationCount' => $read_since_modification_count,
-            'draft' => $itemService->getItem($itemId)->isDraft(),
+            'draft' => $this->itemService->getItem($itemId)->isDraft(),
             'showCategories' => $current_context->withTags(),
             'showHashtags' => $current_context->withBuzzwords(),
-            'language' => $legacyEnvironment->getCurrentContextItem()->getLanguage(),
+            'language' => $this->legacyEnvironment->getCurrentContextItem()->getLanguage(),
             'showAssociations' => $current_context->isAssociationShowExpanded(),
             'buzzExpanded' => $current_context->isBuzzwordShowExpanded(),
             'catzExpanded' => $current_context->isTagsShowExpanded(),
             'roomCategories' => $categories,
-            'isParticipating' => $date->isParticipant($legacyEnvironment->getCurrentUserItem()),
+            'isParticipating' => $date->isParticipant($this->legacyEnvironment->getCurrentUserItem()),
             'isRecurring' => ($date->getRecurrenceId() != ''),
             'alert' => $alert,
             'pathTopicItem' => $pathTopicItem,
@@ -538,16 +526,12 @@ class DateController extends BaseController
     /**
      * @Route("/room/{roomId}/date/events")
      * @param Request $request
-     * @param DateService $dateService
-     * @param TranslatorInterface $translator
      * @param int $roomId
      * @return JsonResponse
      * @throws Exception
      */
     public function eventsAction(
         Request $request,
-        DateService $dateService,
-        TranslatorInterface $translator,
         int $roomId
     ) {
         $roomItem = $this->getRoom($roomId);
@@ -569,7 +553,7 @@ class DateController extends BaseController
             $filterForm->submit($dateFilter);
 
             // set filter conditions on the date manager
-            $dateService->setFilterConditions($filterForm);
+            $this->dateService->setFilterConditions($filterForm);
 
             if (isset($dateFilter['date-from']['date']) && !empty($dateFilter['date-from']['date'])) {
                 $startDate = DateTime::createFromFormat('d.m.Y', $dateFilter['date-from']['date'])->format('Y-m-d 00:00:00');
@@ -578,10 +562,10 @@ class DateController extends BaseController
                 $endDate = DateTime::createFromFormat('d.m.Y', $dateFilter['date-until']['date'])->format('Y-m-d 23:59:59');
             }
         } else {
-            $dateService->setPastFilter(true);
+            $this->dateService->setPastFilter(true);
         }
 
-        $listDates = $dateService->getCalendarEvents($roomId, $startDate, $endDate);
+        $listDates = $this->dateService->getCalendarEvents($roomId, $startDate, $endDate);
 
         $events = array();
         foreach ($listDates as $date) {
@@ -638,35 +622,35 @@ class DateController extends BaseController
                 }
                 
                 if ($recurrencePattern['recurring_select'] == 'RecurringDailyType') {
-                    $recurringDescription = $translator->trans('dailyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDay'], '%date%' => $endDate->format('d.m.Y')), 'date');
+                    $recurringDescription = $this->translator->trans('dailyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDay'], '%date%' => $endDate->format('d.m.Y')), 'date');
                 } else if ($recurrencePattern['recurring_select'] == 'RecurringWeeklyType') {
                     $daysOfWeek = array();
                     if (isset($recurrencePattern['recurring_sub']['recurrenceDaysOfWeek'])) {
                         foreach ($recurrencePattern['recurring_sub']['recurrenceDaysOfWeek'] as $day) {
-                            $daysOfWeek[] = $translator->trans($day, array(), 'date');
+                            $daysOfWeek[] = $this->translator->trans($day, array(), 'date');
                         }
                     }
-                    $recurringDescription = $translator->trans('weeklyDescription', array('%week%' => $recurrencePattern['recurring_sub']['recurrenceWeek'], '%daysOfWeek%' => implode(', ', $daysOfWeek), '%date%' => $endDate->format('d.m.Y')), 'date');
+                    $recurringDescription = $this->translator->trans('weeklyDescription', array('%week%' => $recurrencePattern['recurring_sub']['recurrenceWeek'], '%daysOfWeek%' => implode(', ', $daysOfWeek), '%date%' => $endDate->format('d.m.Y')), 'date');
                 } else if ($recurrencePattern['recurring_select'] == 'RecurringMonthlyType') {
-                    $tempDayOfMonthInterval = $translator->trans('first', array(), 'date');
+                    $tempDayOfMonthInterval = $this->translator->trans('first', array(), 'date');
                     if (isset($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'])) {
                         if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 2) {
-                            $tempDayOfMonthInterval = $translator->trans('second', array(), 'date');
+                            $tempDayOfMonthInterval = $this->translator->trans('second', array(), 'date');
                         } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 3) {
-                            $tempDayOfMonthInterval = $translator->trans('third', array(), 'date');
+                            $tempDayOfMonthInterval = $this->translator->trans('third', array(), 'date');
                         } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 4) {
-                            $tempDayOfMonthInterval = $translator->trans('fourth', array(), 'date');
+                            $tempDayOfMonthInterval = $this->translator->trans('fourth', array(), 'date');
                         } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 5) {
-                            $tempDayOfMonthInterval = $translator->trans('fifth', array(), 'date');
+                            $tempDayOfMonthInterval = $this->translator->trans('fifth', array(), 'date');
                         } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 'last') {
-                            $tempDayOfMonthInterval = $translator->trans('last', array(), 'date');
+                            $tempDayOfMonthInterval = $this->translator->trans('last', array(), 'date');
                         }
                     }
                     if (isset($recurrencePattern['recurring_sub']['recurrenceDayOfMonth'])) {
-                        $recurringDescription = $translator->trans('monthlyDescription', array('%month%' => $recurrencePattern['recurring_sub']['recurrenceMonth'], '%day%' => $tempDayOfMonthInterval, '%dayOfWeek%' => $translator->trans($recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
+                        $recurringDescription = $this->translator->trans('monthlyDescription', array('%month%' => $recurrencePattern['recurring_sub']['recurrenceMonth'], '%day%' => $tempDayOfMonthInterval, '%dayOfWeek%' => $this->translator->trans($recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
                     }
                 } else if ($recurrencePattern['recurring_select'] == 'RecurringYearlyType') {
-                    $recurringDescription = $translator->trans('yearlyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], '%month%' => $translator->trans($recurrencePattern['recurring_sub']['recurrenceMonthOfYear'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
+                    $recurringDescription = $this->translator->trans('yearlyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], '%month%' => $this->translator->trans($recurrencePattern['recurring_sub']['recurrenceMonthOfYear'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
                 }
             }
 
@@ -694,27 +678,19 @@ class DateController extends BaseController
 
     /**
      * @Route("/room/{roomId}/date/eventsdashboard")
-     * @param ItemService $itemService
-     * @param DateService $dateService
-     * @param LegacyEnvironment $environment
      * @return JsonResponse
      * @throws Exception
      */
-    public function eventsdashboardAction(
-        ItemService $itemService,
-        DateService $dateService,
-        LegacyEnvironment $environment
-    ) {
-        $legacyEnvironment = $environment->getEnvironment();
-        
-        $user = $legacyEnvironment->getCurrentUserItem();
+    public function eventsdashboardAction() {
+
+        $user = $this->legacyEnvironment->getCurrentUserItem();
         $userList = $user->getRelatedUserList()->to_array();
 
         $listDates = array();
         foreach ($userList as $tempUser) {
             /** @var cs_user_item $tempUser */
             if ($tempUser->getStatus() >= 2) {
-                $listDates = array_merge($listDates, $dateService->getCalendarEvents($tempUser->getContextId(), $_GET['start'], $_GET['end']));
+                $listDates = array_merge($listDates, $this->dateService->getCalendarEvents($tempUser->getContextId(), $_GET['start'], $_GET['end']));
             }
         }
 
@@ -766,8 +742,6 @@ class DateController extends BaseController
 
             $recurringDescription = '';
             if ($date->getRecurrencePattern() != '') {
-                $translator = $this->get('translator');
-                
                 $recurrencePattern = $date->getRecurrencePattern();
                 
                 if (isset($recurrencePattern['recurringEndDate'])) {
@@ -775,33 +749,33 @@ class DateController extends BaseController
                 }
                 
                 if ($recurrencePattern['recurring_select'] == 'RecurringDailyType') {
-                    $recurringDescription = $translator->trans('dailyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDay'], '%date%' => $endDate->format('d.m.Y')), 'date');
+                    $recurringDescription = $this->translator->trans('dailyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDay'], '%date%' => $endDate->format('d.m.Y')), 'date');
                 } else if ($recurrencePattern['recurring_select'] == 'RecurringWeeklyType') {
                     $daysOfWeek = array();
                     foreach ($recurrencePattern['recurring_sub']['recurrenceDaysOfWeek'] as $day) {
-                        $daysOfWeek[] = $translator->trans($day, array(), 'date');
+                        $daysOfWeek[] = $this->translator->trans($day, array(), 'date');
                     }
-                    $recurringDescription = $translator->trans('weeklyDescription', array('%week%' => $recurrencePattern['recurring_sub']['recurrenceWeek'], '%daysOfWeek%' => implode(', ', $daysOfWeek), '%date%' => $endDate->format('d.m.Y')), 'date');
+                    $recurringDescription = $this->translator->trans('weeklyDescription', array('%week%' => $recurrencePattern['recurring_sub']['recurrenceWeek'], '%daysOfWeek%' => implode(', ', $daysOfWeek), '%date%' => $endDate->format('d.m.Y')), 'date');
                 } else if ($recurrencePattern['recurring_select'] == 'RecurringMonthlyType') {
-                    $tempDayOfMonthInterval = $translator->trans('first', array(), 'date');
+                    $tempDayOfMonthInterval = $this->translator->trans('first', array(), 'date');
                     if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 2) {
-                        $tempDayOfMonthInterval = $translator->trans('second', array(), 'date');
+                        $tempDayOfMonthInterval = $this->translator->trans('second', array(), 'date');
                     } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 3) {
-                        $tempDayOfMonthInterval = $translator->trans('third', array(), 'date');
+                        $tempDayOfMonthInterval = $this->translator->trans('third', array(), 'date');
                     } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 4) {
-                        $tempDayOfMonthInterval = $translator->trans('fourth', array(), 'date');
+                        $tempDayOfMonthInterval = $this->translator->trans('fourth', array(), 'date');
                     } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 5) {
-                        $tempDayOfMonthInterval = $translator->trans('fifth', array(), 'date');
+                        $tempDayOfMonthInterval = $this->translator->trans('fifth', array(), 'date');
                     } else if ($recurrencePattern['recurring_sub']['recurrenceDayOfMonthInterval'] == 'last') {
-                        $tempDayOfMonthInterval = $translator->trans('last', array(), 'date');
+                        $tempDayOfMonthInterval = $this->translator->trans('last', array(), 'date');
                     }
-                    $recurringDescription = $translator->trans('monthlyDescription', array('%month%' => $recurrencePattern['recurring_sub']['recurrenceMonth'], '%day%' => $tempDayOfMonthInterval, '%dayOfWeek%' => $translator->trans($recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
+                    $recurringDescription = $this->translator->trans('monthlyDescription', array('%month%' => $recurrencePattern['recurring_sub']['recurrenceMonth'], '%day%' => $tempDayOfMonthInterval, '%dayOfWeek%' => $this->translator->trans($recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
                 } else if ($recurrencePattern['recurring_select'] == 'RecurringYearlyType') {
-                    $recurringDescription = $translator->trans('yearlyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], '%month%' => $translator->trans($recurrencePattern['recurring_sub']['recurrenceMonthOfYear'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
+                    $recurringDescription = $this->translator->trans('yearlyDescription', array('%day%' => $recurrencePattern['recurring_sub']['recurrenceDayOfMonth'], '%month%' => $this->translator->trans($recurrencePattern['recurring_sub']['recurrenceMonthOfYear'], array(), 'date'), '%date%' => $endDate->format('d.m.Y')), 'date');
                 }
             }
             
-            $context = $itemService->getTypedItem($date->getContextId());
+            $context = $this->itemService->getTypedItem($date->getContextId());
 
             $events[] = array('itemId' => $date->getItemId(),
                               'title' => $date->getTitle(),
@@ -827,19 +801,17 @@ class DateController extends BaseController
 
     /**
      * @Route("/room/{roomId}/date/create/{dateDescription}")
-     * @param DateService $dateService
      * @param $roomId
      * @param $dateDescription
      * @return RedirectResponse
      * @Security("is_granted('ITEM_EDIT', 'NEW') and is_granted('RUBRIC_SEE', 'date')")
      */
     public function createAction(
-        DateService $dateService,
         int $roomId,
         $dateDescription
     ) {
         // create new material item
-        $dateItem = $dateService->getNewDate();
+        $dateItem = $this->dateService->getNewDate();
         $dateItem->setDraftStatus(1);
         $dateItem->setPrivateEditing('1');
 
@@ -885,21 +857,15 @@ class DateController extends BaseController
     /**
      * @Route("/room/{roomId}/date/{itemId}/calendaredit")
      * @param Request $request
-     * @param DateService $dateService
-     * @param TranslatorInterface $translator
      * @param int $itemId
      * @return JsonResponse
      * @throws Exception
      */
     public function calendareditAction(
         Request $request,
-        DateService $dateService,
-        TranslatorInterface $translator,
-        LegacyEnvironment $legacyEnvironment,
         int $itemId
     ) {
-        $legacyEnvironment = $legacyEnvironment->getEnvironment();
-        $date = $dateService->getDate($itemId);
+        $date = $this->dateService->getDate($itemId);
 
         $requestContent = json_decode($request->getContent());
 
@@ -946,11 +912,11 @@ class DateController extends BaseController
         }
 
         // update modifier
-        $date->setModificatorItem($legacyEnvironment->getCurrentUserItem());
+        $date->setModificatorItem($this->legacyEnvironment->getCurrentUserItem());
 
         $date->save();
 
-        $message = '<i class=\'uk-icon-justify uk-icon-medium uk-icon-check-square-o\'></i> ' . $translator->trans('date changed', [], 'date');
+        $message = '<i class=\'uk-icon-justify uk-icon-medium uk-icon-check-square-o\'></i> ' . $this->translator->trans('date changed', [], 'date');
 
         $start = $date->getStartingDay();
         if ($date->getStartingTime() != '') {
@@ -988,13 +954,8 @@ class DateController extends BaseController
      * @Security("is_granted('ITEM_EDIT', itemId) and is_granted('RUBRIC_SEE', 'date')")
      * @param Request $request
      * @param CategoryService $categoryService
-     * @param DateService $dateService
-     * @param ItemService $itemService
      * @param DateTransformer $transformer
-     * @param TranslatorInterface $translator
      * @param ItemController $itemController
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @param int $itemId
      * @return array|RedirectResponse
@@ -1002,20 +963,14 @@ class DateController extends BaseController
     public function editAction(
         Request $request,
         CategoryService $categoryService,
-        DateService $dateService,
-        ItemService $itemService,
         DateTransformer $transformer,
-        TranslatorInterface $translator,
         ItemController $itemController,
-        EventDispatcherInterface $eventDispatcher,
-        LegacyEnvironment $environment,
         int $roomId,
         int $itemId
     ) {
-        $item = $itemService->getItem($itemId);
+        $item = $this->itemService->getItem($itemId);
 
-        $legacyEnvironment = $environment->getEnvironment();
-        $current_context = $legacyEnvironment->getCurrentContextItem();
+        $current_context = $this->legacyEnvironment->getCurrentContextItem();
 
         $isDraft = $item->isDraft();
 
@@ -1023,7 +978,7 @@ class DateController extends BaseController
         $hashtagsMandatory = $current_context->withBuzzwords() && $current_context->isBuzzwordMandatory();
 
         // get date from DateService
-        $dateItem = $dateService->getDate($itemId);
+        $dateItem = $this->dateService->getDate($itemId);
         if (!$dateItem) {
             throw $this->createNotFoundException('No date found for id ' . $itemId);
         }
@@ -1031,9 +986,9 @@ class DateController extends BaseController
         $formData = $transformer->transform($dateItem);
         $formData['categoriesMandatory'] = $categoriesMandatory;
         $formData['hashtagsMandatory'] = $hashtagsMandatory;
-        $formData['language'] = $legacyEnvironment->getCurrentContextItem()->getLanguage();
+        $formData['language'] = $this->legacyEnvironment->getCurrentContextItem()->getLanguage();
         $formData['category_mapping']['categories'] = $itemController->getLinkedCategories($item);
-        $formData['hashtag_mapping']['hashtags'] = $itemController->getLinkedHashtags($itemId, $roomId, $legacyEnvironment);
+        $formData['hashtag_mapping']['hashtags'] = $itemController->getLinkedHashtags($itemId, $roomId, $this->legacyEnvironment);
         $formData['draft'] = $isDraft;
 
         $em = $this->getDoctrine()->getManager();
@@ -1055,15 +1010,15 @@ class DateController extends BaseController
                 'roomId' => $roomId,
                 'itemId' => $itemId,
             )),
-            'placeholderText' => '['.$translator->trans('insert title').']',
+            'placeholderText' => '['.$this->translator->trans('insert title').']',
             'calendars' => $calendarsOptions,
             'calendarsAttr' => $calendarsOptionsAttr,
             'categoryMappingOptions' => [
                 'categories' => $itemController->getCategories($roomId, $categoryService)
             ],
             'hashtagMappingOptions' => [
-                'hashtags' => $itemController->getHashtags($roomId, $legacyEnvironment),
-                'hashTagPlaceholderText' => $translator->trans('Hashtag', [], 'hashtag'),
+                'hashtags' => $itemController->getHashtags($roomId, $this->legacyEnvironment),
+                'hashTagPlaceholderText' => $this->translator->trans('Hashtag', [], 'hashtag'),
                 'hashtagEditUrl' => $this->generateUrl('app_hashtag_add', ['roomId' => $roomId])
             ],
         );
@@ -1087,7 +1042,7 @@ class DateController extends BaseController
                 $dateItem = $transformer->applyTransformation($dateItem, $formData);
 
                 // update modifier
-                $dateItem->setModificatorItem($legacyEnvironment->getCurrentUserItem());
+                $dateItem->setModificatorItem($this->legacyEnvironment->getCurrentUserItem());
 
                 // set linked hashtags and categories
                 $formData = $form->getData();
@@ -1140,17 +1095,17 @@ class DateController extends BaseController
                     $dateItem->setDateTime_recurrence($dateItem->getDateTime_start());
                 }
                 $dateItem = $transformer->applyTransformation($dateItem, $formData);
-                $dateItem->setModificatorItem($legacyEnvironment->getCurrentUserItem());
+                $dateItem->setModificatorItem($this->legacyEnvironment->getCurrentUserItem());
                 $dateItem->save();
             } else if ($saveType == 'saveAllDates') {
-                $datesArray = $dateService->getRecurringDates($dateItem->getContextId(), $dateItem->getRecurrenceId());
+                $datesArray = $this->dateService->getRecurringDates($dateItem->getContextId(), $dateItem->getRecurrenceId());
                 $dateItem = $transformer->applyTransformation($dateItem, $formData);
-                $dateItem->setModificatorItem($legacyEnvironment->getCurrentUserItem());
+                $dateItem->setModificatorItem($this->legacyEnvironment->getCurrentUserItem());
                 $dateItem->save();
                 foreach ($datesArray as $tempDate) {
                     $tempDate->setTitle($dateItem->getTitle());
                     $tempDate->setPublic((int)$dateItem->isPublic());
-                    $tempDate->setModificatorItem($legacyEnvironment->getCurrentUserItem());
+                    $tempDate->setModificatorItem($this->legacyEnvironment->getCurrentUserItem());
                     $tempDate->setColor($dateItem->getColor());
                     $tempDate->setCalendarId($dateItem->getCalendarId());
                     $tempDate->setWholeDay($dateItem->isWholeDay());
@@ -1160,25 +1115,25 @@ class DateController extends BaseController
                     $tempDate->save();
                     
                     // mark as read and noticed by creator
-                    $reader_manager = $legacyEnvironment->getReaderManager();
+                    $reader_manager = $this->legacyEnvironment->getReaderManager();
                     $reader_manager->markRead($tempDate->getItemID(), $tempDate->getVersionID());
 
-                    $noticed_manager = $legacyEnvironment->getNoticedManager();
+                    $noticed_manager = $this->legacyEnvironment->getNoticedManager();
                     $noticed_manager->markNoticed($tempDate->getItemID(), $tempDate->getVersionID());
                 }
             }
             return $this->redirectToRoute('app_date_save', array('roomId' => $roomId, 'itemId' => $itemId));
         }
 
-        $eventDispatcher->dispatch(new CommsyEditEvent($dateItem), 'commsy.edit');
+        $this->eventDispatcher->dispatch(new CommsyEditEvent($dateItem), CommsyEditEvent::EDIT);
 
         return array(
             'form' => $form->createView(),
             'isDraft' => $isDraft,
             'showHashtags' => $hashtagsMandatory,
-            'language' => $legacyEnvironment->getCurrentContextItem()->getLanguage(),
+            'language' => $this->legacyEnvironment->getCurrentContextItem()->getLanguage(),
             'showCategories' => $categoriesMandatory,
-            'currentUser' => $legacyEnvironment->getCurrentUserItem(),
+            'currentUser' => $this->legacyEnvironment->getCurrentUserItem(),
             'withRecurrence' => $dateItem->getRecurrencePattern() != '',
             'date' => $dateItem,
         );
@@ -1223,37 +1178,26 @@ class DateController extends BaseController
      * @Route("/room/{roomId}/date/{itemId}/save")
      * @Template()
      * @Security("is_granted('ITEM_EDIT', itemId) and is_granted('RUBRIC_SEE', 'date')")
-     * @param ItemService $itemService
-     * @param DateService $dateService
-     * @param ReaderService $readerService
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @param int $itemId
      * @return array
      */
     public function saveAction(
-        ItemService $itemService,
-        DateService $dateService,
-        ReaderService $readerService,
-        EventDispatcherInterface $eventDispatcher,
-        LegacyEnvironment $environment,
         int $roomId,
         int $itemId
     ) {
-        $date = $dateService->getDate($itemId);
+        $date = $this->dateService->getDate($itemId);
         
         $itemArray = array($date);
         $modifierList = array();
         foreach ($itemArray as $item) {
-            $modifierList[$item->getItemId()] = $itemService->getAdditionalEditorsForItem($item);
+            $modifierList[$item->getItemId()] = $this->itemService->getAdditionalEditorsForItem($item);
         }
         
-        $legacyEnvironment = $environment->getEnvironment();
-        $readerManager = $legacyEnvironment->getReaderManager();
+        $readerManager = $this->legacyEnvironment->getReaderManager();
         
-        $userManager = $legacyEnvironment->getUserManager();
-        $userManager->setContextLimit($legacyEnvironment->getCurrentContextID());
+        $userManager = $this->legacyEnvironment->getUserManager();
+        $userManager->setContextLimit($this->legacyEnvironment->getCurrentContextID());
         $userManager->setUserLimit();
         $userManager->select();
         $user_list = $userManager->get();
@@ -1286,17 +1230,17 @@ class DateController extends BaseController
         $readerList = array();
         $modifierList = array();
         foreach ($itemArray as $item) {
-            $reader = $readerService->getLatestReader($item->getItemId());
+            $reader = $this->readerService->getLatestReader($item->getItemId());
             if ( empty($reader) ) {
                $readerList[$item->getItemId()] = 'new';
             } elseif ( $reader['read_date'] < $item->getModificationDate() ) {
                $readerList[$item->getItemId()] = 'changed';
             }
             
-            $modifierList[$item->getItemId()] = $itemService->getAdditionalEditorsForItem($item);
+            $modifierList[$item->getItemId()] = $this->itemService->getAdditionalEditorsForItem($item);
         }
 
-        $eventDispatcher->dispatch(new CommsyEditEvent($date), 'commsy.save');
+        $this->eventDispatcher->dispatch(new CommsyEditEvent($date),CommsyEditEvent::SAVE);
 
         return array(
             'roomId' => $roomId,
@@ -1310,8 +1254,6 @@ class DateController extends BaseController
     
     function saveRecurringDates($dateItem, $isNewRecurring, $valuesToChange, $formData)
     {
-        $legacyEnvironment = $this->get('commsy_legacy.environment')->getEnvironment();
-
         /** @var cs_dates_item $dateItem */
 
         if($isNewRecurring) {
@@ -1509,10 +1451,10 @@ class DateController extends BaseController
                 $tempDate->save();
 
                 // mark as read and noticed by creator
-                $reader_manager = $legacyEnvironment->getReaderManager();
+                $reader_manager = $this->legacyEnvironment->getReaderManager();
                 $reader_manager->markRead($tempDate->getItemID(), $tempDate->getVersionID());
 
-                $noticed_manager = $legacyEnvironment->getNoticedManager();
+                $noticed_manager = $this->legacyEnvironment->getNoticedManager();
                 $noticed_manager->markNoticed($tempDate->getItemID(), $tempDate->getVersionID());
 
             }
@@ -1521,7 +1463,7 @@ class DateController extends BaseController
             $dateItem->save();
         } else {
             // TODO: remove this else block if (as suspected) it is dead code that doesn't get executed anymore
-            $datesManager = $legacyEnvironment->getDatesManager();
+            $datesManager = $this->legacyEnvironment->getDatesManager();
             $datesManager->resetLimits();
             $datesManager->setRecurrenceLimit($dateItem->getRecurrenceId());
             $datesManager->setWithoutDateModeLimit();
@@ -1549,10 +1491,10 @@ class DateController extends BaseController
                 }
 
                 // mark as read and noticed by creator
-                $reader_manager = $legacyEnvironment->getReaderManager();
+                $reader_manager = $this->legacyEnvironment->getReaderManager();
                 $reader_manager->markRead($tempDate->getItemID(), $tempDate->getVersionID());
 
-                $noticed_manager = $legacyEnvironment->getNoticedManager();
+                $noticed_manager = $this->legacyEnvironment->getNoticedManager();
                 $noticed_manager->markNoticed($tempDate->getItemID(), $tempDate->getVersionID());
 
                 //$tempDate->save();
@@ -1564,35 +1506,26 @@ class DateController extends BaseController
     /**
      * @Route("/room/{roomId}/date/{itemId}/print")
      * @param CategoryService $categoryService
-     * @param DateService $dateService
-     * @param ItemService $itemService
      * @param PrintService $printService
-     * @param ReaderService $readerService
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @param int $itemId
      * @return Response
      */
     public function printAction(
         CategoryService $categoryService,
-        DateService $dateService,
-        ItemService $itemService,
         PrintService $printService,
-        ReaderService $readerService,
-        LegacyEnvironment $environment,
         int $roomId,
         int $itemId
     ) {
-        $date = $dateService->getDate($itemId);
-        $legacyEnvironment = $environment->getEnvironment();
+        $date = $this->dateService->getDate($itemId);
         $item = $date;
-        $reader_manager = $legacyEnvironment->getReaderManager();
+        $reader_manager = $this->legacyEnvironment->getReaderManager();
         $reader = $reader_manager->getLatestReader($item->getItemID());
         if(empty($reader) || $reader['read_date'] < $item->getModificationDate()) {
             $reader_manager->markRead($item->getItemID(), $item->getVersionID());
         }
 
-        $noticed_manager = $legacyEnvironment->getNoticedManager();
+        $noticed_manager = $this->legacyEnvironment->getNoticedManager();
         $noticed = $noticed_manager->getLatestNoticed($item->getItemID());
         if(empty($noticed) || $noticed['read_date'] < $item->getModificationDate()) {
             $noticed_manager->markNoticed($item->getItemID(), $item->getVersionID());
@@ -1601,12 +1534,12 @@ class DateController extends BaseController
         
         $itemArray = array($date);
 
-        $current_context = $legacyEnvironment->getCurrentContextItem();
+        $current_context = $this->legacyEnvironment->getCurrentContextItem();
 
-        $readerManager = $legacyEnvironment->getReaderManager();
+        $readerManager = $this->legacyEnvironment->getReaderManager();
 
-        $userManager = $legacyEnvironment->getUserManager();
-        $userManager->setContextLimit($legacyEnvironment->getCurrentContextID());
+        $userManager = $this->legacyEnvironment->getUserManager();
+        $userManager->setContextLimit($this->legacyEnvironment->getCurrentContextID());
         $userManager->setUserLimit();
         $userManager->select();
         $user_list = $userManager->get();
@@ -1640,14 +1573,14 @@ class DateController extends BaseController
         $readerList = array();
         $modifierList = array();
         foreach ($itemArray as $item) {
-            $reader = $readerService->getLatestReader($item->getItemId());
+            $reader = $this->readerService->getLatestReader($item->getItemId());
             if ( empty($reader) ) {
                $readerList[$item->getItemId()] = 'new';
             } elseif ( $reader['read_date'] < $item->getModificationDate() ) {
                $readerList[$item->getItemId()] = 'changed';
             }
             
-            $modifierList[$item->getItemId()] = $itemService->getAdditionalEditorsForItem($item);
+            $modifierList[$item->getItemId()] = $this->itemService->getAdditionalEditorsForItem($item);
         }
 
         // annotation form
@@ -1662,19 +1595,19 @@ class DateController extends BaseController
 
         $html = $this->renderView('date/detail_print.html.twig', [
             'roomId' => $roomId,
-            'date' => $dateService->getDate($itemId),
+            'date' => $this->dateService->getDate($itemId),
             'readerList' => $readerList,
             'modifierList' => $modifierList,
-            'user' => $legacyEnvironment->getCurrentUserItem(),
+            'user' => $this->legacyEnvironment->getCurrentUserItem(),
             'annotationForm' => $form->createView(),
             'userCount' => $all_user_count,
             'readCount' => $read_count,
             'readSinceModificationCount' => $read_since_modification_count,
-            'draft' => $itemService->getItem($itemId)->isDraft(),
+            'draft' => $this->itemService->getItem($itemId)->isDraft(),
             'showCategories' => $current_context->withTags(),
             'showAssociations' => $current_context->isAssociationShowExpanded(),
             'showHashtags' => $current_context->withBuzzwords(),
-            'language' => $legacyEnvironment->getCurrentContextItem()->getLanguage(),
+            'language' => $this->legacyEnvironment->getCurrentContextItem()->getLanguage(),
             'buzzExpanded' => $current_context->isBuzzwordShowExpanded(),
             'catzExpanded' => $current_context->isTagsShowExpanded(),
             'roomCategories' => $categories,
@@ -1685,25 +1618,20 @@ class DateController extends BaseController
 
     /**
      * @Route("/room/{roomId}/date/{itemId}/participate")
-     * @param DateService $dateService
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @param int $itemId
      * @return RedirectResponse
      * @Security("is_granted('ITEM_PARTICIPATE', itemId) and is_granted('RUBRIC_SEE', 'date')")
      */
     public function participateAction(
-        DateService $dateService,
-        LegacyEnvironment $environment,
         int $roomId,
         int $itemId
     ) {
-        $date = $dateService->getDate($itemId);
+        $date = $this->dateService->getDate($itemId);
 
-        $legacyEnvironment = $environment->getEnvironment();
-        $currentUser = $legacyEnvironment->getCurrentUserItem();
+        $currentUser = $this->legacyEnvironment->getCurrentUserItem();
         
-        if (!$date->isParticipant($legacyEnvironment->getCurrentUserItem())) {
+        if (!$date->isParticipant($this->legacyEnvironment->getCurrentUserItem())) {
             $date->addParticipant($currentUser);
         } else {
             $date->removeParticipant($currentUser);
@@ -1717,26 +1645,20 @@ class DateController extends BaseController
      * @Template()
      * @param Request $request
      * @param CalendarsService $calendarsService
-     * @param TranslatorInterface $translator
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @return array|RedirectResponse
      */
     public function importAction(
         Request $request,
         CalendarsService $calendarsService,
-        TranslatorInterface $translator,
-        LegacyEnvironment $environment,
         int $roomId
     ) {
-        $legacyEnvironment = $environment->getEnvironment();
-
         $formData = [];
         $em = $this->getDoctrine()->getManager();
         $repository = $em->getRepository('App:Calendars');
         $calendars = $repository->findBy(array('context_id' => $roomId));
-        $calendarsOptions = [$translator->trans('new calendar', [], 'date') => 'new'];
-        $calendarsOptionsAttr = [['title' => $translator->trans('new calendar'), 'color' => '#ffffff', 'hasLightColor' => true]];
+        $calendarsOptions = [$this->translator->trans('new calendar', [], 'date') => 'new'];
+        $calendarsOptionsAttr = [['title' => $this->translator->trans('new calendar'), 'color' => '#ffffff', 'hasLightColor' => true]];
         foreach ($calendars as $calendar) {
             if (!$calendar->getExternalUrl()) {
                 $calendarsOptions[$calendar->getTitle()] = $calendar->getId();
@@ -1779,12 +1701,12 @@ class DateController extends BaseController
 
                     $calendarTitle = $formData['calendartitle'];
                     if ($calendarTitle == '') {
-                        $calendarTitle = $translator->trans('new calendar');
+                        $calendarTitle = $this->translator->trans('new calendar');
                     }
                     $calendar->setTitle($calendarTitle);
 
                     $calendar->setContextId($roomId);
-                    $calendar->setCreatorId($legacyEnvironment->getCurrentUserId());
+                    $calendar->setCreatorId($this->legacyEnvironment->getCurrentUserId());
 
                     $calendarColor = $formData['calendarcolor'];
                     if ($calendarColor == '') {
@@ -1918,6 +1840,8 @@ class DateController extends BaseController
      */
     public function xhrDeleteAction(
         Request $request,
+        DeleteAction $action,
+        DeleteDate $deleteDate,
         int $roomId)
     {
         $room = $this->getRoom($roomId);
@@ -1930,14 +1854,13 @@ class DateController extends BaseController
             $recurring = isset($payload['recurring']) ?? false;
         }
 
-        $action = $this->get('commsy.action.delete.date');
         $strategy = $action->getStrategy();
 
         if ($strategy instanceof DeleteDate) {
             $strategy->setRecurring($recurring);
             $strategy->setDateMode($room->getDatesPresentationStatus());
         }
-
+        $action->setDeleteStrategy($deleteDate);
         return $action->execute($room, $items);
     }
 
@@ -1976,8 +1899,6 @@ class DateController extends BaseController
         $selectAll,
         $itemIds = [])
     {
-        $dateService = $this->get('commsy_legacy.date_service');
-
         if ($selectAll) {
             if ($request->query->has('date_filter')) {
                 $currentFilter = $request->query->get('date_filter');
@@ -1987,15 +1908,15 @@ class DateController extends BaseController
                 $filterForm->submit($currentFilter);
 
                 // apply filter
-                $dateService->setFilterConditions($filterForm);
+                $this->dateService->setFilterConditions($filterForm);
             } else {
-                $dateService->setPastFilter(false);
-                $dateService->hideDeactivatedEntries();
+                $this->dateService->setPastFilter(false);
+                $this->dateService->hideDeactivatedEntries();
             }
 
-            return $dateService->getListDates($roomItem->getItemID());
+            return $this->dateService->getListDates($roomItem->getItemID());
         } else {
-            return $dateService->getDatesById($roomItem->getItemID(), $itemIds);
+            return $this->dateService->getDatesById($roomItem->getItemID(), $itemIds);
         }
     }
 }
