@@ -28,6 +28,49 @@ use Symfony\Component\HttpFoundation\Request;
  */
 class PortfolioController extends AbstractController
 {
+
+    private PortfolioService $portfolioService;
+
+    private PortfolioTransformer $transformer;
+
+    private ReaderService $readerService;
+
+    private ItemService $itemService;
+
+    private UserService $userService;
+
+    private CategoryService $categoryService;
+
+    private \cs_environment $legacyEnvironment;
+
+    /**
+     * PortfolioController constructor.
+     * @param PortfolioService $portfolioService
+     * @param PortfolioTransformer $transformer
+     * @param ReaderService $readerService
+     * @param ItemService $itemService
+     * @param UserService $userService
+     * @param CategoryService $categoryService
+     * @param LegacyEnvironment $legacyEnvironment
+     */
+    public function __construct(PortfolioService $portfolioService,
+                                PortfolioTransformer $transformer,
+                                ReaderService $readerService,
+                                ItemService $itemService,
+                                UserService $userService,
+                                CategoryService $categoryService,
+                                LegacyEnvironment $legacyEnvironment)
+    {
+        $this->portfolioService = $portfolioService;
+        $this->transformer = $transformer;
+        $this->readerService = $readerService;
+        $this->itemService = $itemService;
+        $this->userService = $userService;
+        $this->categoryService = $categoryService;
+        $this->legacyEnvironment = $legacyEnvironment->getEnvironment();
+    }
+
+
     /**
      * @Route("/room/{roomId}/portfolio/")
      * @Template()
@@ -55,19 +98,15 @@ class PortfolioController extends AbstractController
      *     "portfolioId": "\d+"
      * }))
      * @Template()
-     * @param PortfolioService $portfolioService
-     * @param UserService $userService
      * @param int $roomId
      * @param int|null $portfolioId
      * @return array
      */
     public function portfolioAction(
-        PortfolioService $portfolioService,
-        UserService $userService,
         int $roomId,
         int $portfolioId = null
     ) {
-        $portfolio = $portfolioService->getPortfolio($portfolioId);
+        $portfolio = $this->portfolioService->getPortfolio($portfolioId);
 
         $linkItemIds = [];
         foreach ($portfolio['links'] as $linkArray) {
@@ -117,7 +156,7 @@ class PortfolioController extends AbstractController
         }
 
         /** @var cs_user_item $user */
-        $user = $userService->getPortalUserFromSessionId();
+        $user = $this->userService->getPortalUserFromSessionId();
 
         $external = false;
         if ($user->getRelatedPrivateRoomUserItem()->getItemId() != $portfolio['creatorId']) {
@@ -137,17 +176,15 @@ class PortfolioController extends AbstractController
     /**
      * @Route("/room/{roomId}/portfolio/portfoliosource/{source}")
      * @Template()
-     * @param PortfolioService $portfolioService
      * @param int $roomId
      * @param string|null $source
      * @return array
      */
     public function tabsAction(
-        PortfolioService $portfolioService,
         int $roomId,
         string $source = null
     ) {
-        $portfolioList = $portfolioService->getPortfolioList();
+        $portfolioList = $this->portfolioService->getPortfolioList();
 
         $portfolios = [];
         $myPortfolios = true;
@@ -168,11 +205,6 @@ class PortfolioController extends AbstractController
     /**
      * @Route("/room/{roomId}/portfolio/{portfolioId}/detail/{firstTagId}/{secondTagId}")
      * @Template()
-     * @param CategoryService $categoryService
-     * @param ItemService $itemService
-     * @param PortfolioService $portfolioService
-     * @param ReaderService $readerService
-     * @param UserService $userService
      * @param int $roomId
      * @param int $portfolioId
      * @param int $firstTagId
@@ -180,17 +212,12 @@ class PortfolioController extends AbstractController
      * @return array
      */
     public function detailAction(
-        CategoryService $categoryService,
-        ItemService $itemService,
-        PortfolioService $portfolioService,
-        ReaderService $readerService,
-        UserService $userService,
         int $roomId,
         int $portfolioId,
         int $firstTagId,
         int $secondTagId
     ) {
-        $portfolio = $portfolioService->getPortfolio($portfolioId);
+        $portfolio = $this->portfolioService->getPortfolio($portfolioId);
 
         $items = [];
         foreach ($portfolio['links'] as $tempFirstTagId => $firstEntries) {
@@ -199,7 +226,7 @@ class PortfolioController extends AbstractController
                     foreach ($firstEntries as $firstEntry) {
                         foreach ($secondEntries as $secondEntry) {
                             if ($firstEntry['itemId'] == $secondEntry['itemId']) {
-                                $items[] = $itemService->getTypedItem($firstEntry['itemId']);
+                                $items[] = $this->itemService->getTypedItem($firstEntry['itemId']);
                             }
                         }
                     }
@@ -208,14 +235,14 @@ class PortfolioController extends AbstractController
         }
 
         /** @var cs_user_item $user */
-        $user = $userService->getPortalUserFromSessionId();
+        $user = $this->userService->getPortalUserFromSessionId();
 
         $readerList = array();
         foreach ($items as $item) {
             if ($item != null) {
                 $relatedUser = $user->getRelatedUserItemInContext($item->getContextId());
                 if ($relatedUser) {
-                    $readerList[$item->getItemId()] = $readerService->getChangeStatusForUserByID($item->getItemId(), $relatedUser->getItemId());
+                    $readerList[$item->getItemId()] = $this->readerService->getChangeStatusForUserByID($item->getItemId(), $relatedUser->getItemId());
                 }
             }
         }
@@ -228,8 +255,8 @@ class PortfolioController extends AbstractController
             'items' => $items,
             'feedList' => $items,
             'readerList' => $readerList,
-            'firstTag' => $categoryService->getTag($firstTagId),
-            'secondTag' => $categoryService->getTag($secondTagId),
+            'firstTag' => $this->categoryService->getTag($firstTagId),
+            'secondTag' => $this->categoryService->getTag($secondTagId),
             'annotationForm' => $form->createView(),
             'portfolio' => $portfolio,
             'portfolioId' => $portfolioId,
@@ -242,27 +269,19 @@ class PortfolioController extends AbstractController
      * @Route("/room/{roomId}/portfolio/{portfolioId}/edit")
      * @Template()
      * @param Request $request
-     * @param ItemService $itemService
-     * @param PortfolioService $portfolioService
-     * @param PortfolioTransformer $transformer
-     * @param LegacyEnvironment $legacyEnvironment
      * @param int $roomId
-     * @param int $portfolioId
+     * @param string $portfolioId
      * @return array|RedirectResponse
      */
     public function editAction(
         Request $request,
-        ItemService $itemService,
-        PortfolioService $portfolioService,
-        PortfolioTransformer $transformer,
-        LegacyEnvironment $legacyEnvironment,
         int $roomId,
-        int $portfolioId
+        string $portfolioId
     ) {
 
         // when creating a new item, return a redirect to the edit form (portfolio draft)
         if ($portfolioId === 'new') {
-            $portfolioItem = $portfolioService->getNewItem();
+            $portfolioItem = $this->portfolioService->getNewItem();
             $portfolioItem->save();
             return $this->redirectToRoute('app_portfolio_edit', [
                 'roomId' => $roomId,
@@ -270,12 +289,12 @@ class PortfolioController extends AbstractController
             ]);
         }
 
-        $item = $itemService->getItem($portfolioId);
+        $item = $this->itemService->getItem($portfolioId);
 
-        $portfolioManager = $legacyEnvironment->getEnvironment()->getPortfolioManager();
+        $portfolioManager = $this->legacyEnvironment->getPortfolioManager();
         $portfolioItem = $portfolioManager->getItem($portfolioId);
 
-        $formData = $transformer->transform($portfolioItem);
+        $formData = $this->transformer->transform($portfolioItem);
 
         $form = $this->createForm(PortfolioType::class, $formData, [
             'item' => $item,
@@ -284,7 +303,7 @@ class PortfolioController extends AbstractController
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             if ($form->get('save')->isClicked()) {
-                $portfolioItem = $transformer->applyTransformation($portfolioItem, $form->getData());
+                $portfolioItem = $this->transformer->applyTransformation($portfolioItem, $form->getData());
                 $portfolioItem->save();
 
                 // ensure portfolio is now longer in draft state after saving
@@ -295,7 +314,7 @@ class PortfolioController extends AbstractController
                     $formData = $form->getData();
                     if (isset($formData['from_template'])) {
                         if ($formData['from_template'] != 'none') {
-                            $portfolioService->prepareFromTemplate((int) $formData['from_template'], $portfolioItem);
+                            $this->portfolioService->prepareFromTemplate((int) $formData['from_template'], $portfolioItem);
                         }
                     }
                 }
@@ -327,7 +346,6 @@ class PortfolioController extends AbstractController
      * @Route("/room/{roomId}/portfolio/{portfolioId}/editcategory/{position}/{categoryId}/")
      * @Template()
      * @param Request $request
-     * @param CategoryService $categoryService
      * @param int $roomId
      * @param int $portfolioId
      * @param string $position
@@ -336,7 +354,6 @@ class PortfolioController extends AbstractController
      */
     public function editcategoryAction(
         Request $request,
-        CategoryService $categoryService,
         int $roomId,
         int $portfolioId,
         string $position,
@@ -353,7 +370,7 @@ class PortfolioController extends AbstractController
             $formData['categories'] = [$categoryId];
         }
 
-        $roomTags = $categoryService->getTags($roomId);
+        $roomTags = $this->categoryService->getTags($roomId);
 //        $disabledCategories = $this->getDisabledTags($roomTags, $categoryId, $portfolio);
 
         $form = $this->createForm(PortfolioEditCategoryType::class, $formData, array(
@@ -440,18 +457,15 @@ class PortfolioController extends AbstractController
 
     /**
      * @Route("/room/{roomId}/portfolio/{portfolioId}/stopActivation")
-     * @param LegacyEnvironment $environment
      * @param int $roomId
      * @param int $portfolioId
      * @return RedirectResponse
      */
     public function stopActivation(
-        LegacyEnvironment $environment,
         int $roomId,
         int $portfolioId
     ) {
-        $legacyEnvironment = $environment->getEnvironment();
-        $portfolioManager = $legacyEnvironment->getPortfolioManager();
+        $portfolioManager = $this->legacyEnvironment->getPortfolioManager();
 
         $portfolio = $portfolioManager->getItem($portfolioId);
 
@@ -459,7 +473,7 @@ class PortfolioController extends AbstractController
             throw new NotFoundHttpException('Portfolio not found');
         }
 
-        $currentUser = $legacyEnvironment->getCurrentUserItem();
+        $currentUser = $this->legacyEnvironment->getCurrentUserItem();
 
         $portfolioManager->removeExternalViewer($portfolio->getItemID(), $currentUser->getUserID());
 
