@@ -5,8 +5,8 @@ namespace App\Facade;
 
 
 use App\Entity\Account;
-use App\Entity\User;
 use App\Services\LegacyEnvironment;
+use cs_user_item;
 use Doctrine\ORM\EntityManagerInterface;
 
 class AccountCreatorFacade
@@ -21,7 +21,11 @@ class AccountCreatorFacade
         $this->legacyEnvironment = $legacyEnvironment->getEnvironment();
     }
 
-    public function persistNewAccount(Account $account)
+    /**
+     * @param Account $account
+     * @return cs_user_item
+     */
+    public function persistNewAccount(Account $account): cs_user_item
     {
         $this->entityManager->persist($account);
         $this->entityManager->flush();
@@ -47,5 +51,7 @@ class AccountCreatorFacade
         $portalUser->setLanguage($account->getLanguage());
         $portalUser->makeUser();
         $portalUser->save();
+
+        return $portalUser;
     }
 }
