@@ -4,6 +4,8 @@ namespace App\Repository;
 
 use App\Entity\RoomPrivat;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\ORM\NonUniqueResultException;
+use Doctrine\ORM\NoResultException;
 use Doctrine\ORM\Query\Expr;
 use Doctrine\Persistence\ManagerRegistry;
 
@@ -14,6 +16,12 @@ class RoomPrivateRepository extends ServiceEntityRepository
         parent::__construct($registry, RoomPrivat::class);
     }
 
+    /**
+     * @param int $contextId
+     * @param string $username
+     * @return RoomPrivat|null
+     * @throws NoResultException|NonUniqueResultException
+     */
     public function findByContextIdAndUsername(int $contextId, string $username):? RoomPrivat
     {
         return $this->createQueryBuilder('rp')
@@ -31,6 +39,6 @@ class RoomPrivateRepository extends ServiceEntityRepository
             ])
             ->getQuery()
             ->setMaxResults(1)
-            ->getOneOrNullResult();
+            ->getSingleResult();
     }
 }
